fix(web): wrap useSearchParams in dedicated Suspense boundary
Extract the useSearchParams + auto-launch logic from IssueActionSheet
into a new AutoLaunchTrigger component with its own Suspense boundary.
This prevents useSearchParams from de-opting SSR for the entire route.

Closes #316
