+++ packages/web/components/detail/IssueActionSheet.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { useState, useEffect, useRef, useTransition } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
+import { Suspense, useState, useEffect, useTransition } from "react";
+import { useRouter } from "next/navigation";
 import type {
   GitHubIssue,
   GitHubComment,
@@ -20,6 +20,7 @@ import { useToast } from "@/components/ui/ToastProvider";
 import { newIdempotencyKey } from "@/lib/idempotency-key";
 import { useOfflineAware } from "@/hooks/useOfflineAware";
 import { useStaleTab } from "@/hooks/useStaleTab";
+import { AutoLaunchTrigger } from "./AutoLaunchTrigger";
 import styles from "./ActionSheet.module.css";
 import assignStyles from "../list/AssignSheet.module.css";
 
@@ -67,24 +68,6 @@ export function IssueActionSheet({
   const [reassignKey, setReassignKey] = useState<string | null>(null);
 
   const { isOffline } = useOfflineAware();
-  const searchParams = useSearchParams();
-
-  // Keep a ref to hasLiveDeployment so the mount-only effect below always
-  // reads the latest value without needing it in the dependency array
-  // (which would re-trigger the effect on every render where it changes).
-  const hasLiveDeploymentRef = useRef(hasLiveDeployment);
-  useEffect(() => {
-    hasLiveDeploymentRef.current = hasLiveDeployment;
-  }, [hasLiveDeployment]);
-
-  useEffect(() => {
-    if (searchParams.get("launch") === "true" && !hasLiveDeploymentRef.current) {
-      handleLaunchTap();
-      const url = new URL(window.location.href);
-      url.searchParams.delete("launch");
-      window.history.replaceState({}, "", url.toString());
-    }
-  }, []); // only run on mount
 
   useStaleTab();
 
@@ -232,6 +215,13 @@ export function IssueActionSheet({
         onTrigger={() => setSheetOpen(true)}
         label="Actions"
       />
+
+      <Suspense fallback={null}>
+        <AutoLaunchTrigger
+          hasLiveDeployment={hasLiveDeployment}
+          onTrigger={handleLaunchTap}
+        />
+      </Suspense>
 
       {/* Desktop inline action bar — visible only on wide viewports */}
       <div className={styles.desktopBar}>
